requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-whois==0.9.6
requests==2.32.5
six==1.17.0
urllib3==2.5.0
